Refactor middleware (#325)
* Split middleware into a subpackage

* Refactor basic auth header building

* Add encoding parameter to to_bytes()
